requirements.txt
qrcode==7.3.1
image==1.5.33
python-telegram-bot==13.0
python-dotenv==0.21.0